added support for passenger pickup and full state space
this is now almost solving the full taxi problem.  It is handling all 500 states and handles picking up and dropping off the passenger, but it only has 5 actions with a single action for pickup/dropoff.  I hacked the env for now to automatically select pickup when pass not in taxi and automatically select dropoff when pass in taxi.  I also changed reward structure to give +20 for correct pickups  so that the agent will learn to do correct pickups.  As a result, I added a -24 reward penalty for incorrect dropoff after pickup, so that agent won't end up learning to go into a pickup->dropoff->pickup->dropoff infinite loop to maximize reward without accomplishing objective.